+++ conjurapi/authn.go
@@ -104,7 +104,7 @@ func (c *Client) createAuthRequest(req *http.Request) error {
 
 func (c *Client) ChangeUserPassword(username string, password string, newPassword string) ([]byte, error) {
 	if isConjurCloudURL(c.config.ApplianceURL) {
-		return nil, errors.New("Change User Password is not supported in Conjur Cloud.")
+		return nil, errors.New("Change User Password is not supported in Conjur Cloud")
 	}
 
 	req, err := c.ChangeUserPasswordRequest(username, password, newPassword)
@@ -132,7 +132,7 @@ func (c *Client) ChangeCurrentUserPassword(newPassword string) ([]byte, error) {
 // Login exchanges a user's password for an API key.
 func (c *Client) Login(login string, password string) ([]byte, error) {
 	if isConjurCloudURL(c.config.ApplianceURL) && !strings.HasPrefix(login, "host/") {
-		return nil, errors.New("Login for users is not supported in Conjur Cloud.")
+		return nil, errors.New("Login for users is not supported in Conjur Cloud")
 	}
 
 	req, err := c.LoginRequest(login, password)
@@ -310,7 +310,7 @@ func (c *Client) JWTAuthenticate(jwt, hostID string) ([]byte, error) {
 
 func (c *Client) ListOidcProviders() ([]OidcProvider, error) {
 	if isConjurCloudURL(c.config.ApplianceURL) {
-		return nil, errors.New("List OIDC Providers is not supported in Conjur Cloud.")
+		return nil, errors.New("List OIDC Providers is not supported in Conjur Cloud")
 	}
 
 	req, err := c.ListOidcProvidersRequest()
@@ -359,7 +359,7 @@ func (c *Client) RotateCurrentRoleAPIKey() ([]byte, error) {
 	}
 
 	if isConjurCloudURL(c.config.ApplianceURL) && !strings.HasPrefix(roleID, "host/") {
-		return nil, errors.New("Rotate API Key for users is not supported in Conjur Cloud.")
+		return nil, errors.New("Rotate API Key for users is not supported in Conjur Cloud")
 	}
 
 	resp, err := c.rotateCurrentRoleAPIKey(roleID, password)
@@ -379,7 +379,7 @@ func (c *Client) RotateCurrentRoleAPIKey() ([]byte, error) {
 // The authenticated user must have update privilege on the role.
 func (c *Client) RotateUserAPIKey(userID string) ([]byte, error) {
 	if isConjurCloudURL(c.config.ApplianceURL) {
-		return nil, errors.New("Rotate API Key for users is not supported in Conjur Cloud.")
+		return nil, errors.New("Rotate API Key for users is not supported in Conjur Cloud")
 	}
 	return c.rotateApiKeyAndEnforceKind(userID, "user")
 }
@@ -438,7 +438,7 @@ func (c *Client) rotateCurrentRoleAPIKey(roleID string, password string) (*http.
 
 func (c *Client) PublicKeys(kind string, identifier string) ([]byte, error) {
 	if isConjurCloudURL(c.config.ApplianceURL) {
-		return nil, errors.New("Public Keys is not supported in Conjur Cloud.")
+		return nil, errors.New("Public Keys is not supported in Conjur Cloud")
 	}
 
 	req, err := c.PublicKeysRequest(kind, identifier)