+++ conjurapi/policy.go
@@ -1,6 +1,7 @@
 package conjurapi
 
 import (
+	"errors"
 	"fmt"
 	"io"
 
@@ -84,9 +85,12 @@ func (c *Client) LoadPolicy(mode PolicyMode, policyID string, policy io.Reader) 
 }
 
 func (c *Client) DryRunPolicy(mode PolicyMode, policyID string, policy io.Reader) (*DryRunPolicyResponse, error) {
+	if isConjurCloudURL(c.config.ApplianceURL) {
+		return nil, errors.New("Policy Dry Run is not supported in Conjur Cloud")
+	}
 	err := c.VerifyMinServerVersion("1.21.1")
 	if err != nil {
-		return nil, fmt.Errorf("Dry run policy is not supported in Conjur versions older than 1.21.1")
+		return nil, fmt.Errorf("Policy Dry Run is not supported in Conjur versions older than 1.21.1")
 	}
 
 	req, err := c.LoadPolicyRequest(mode, policyID, policy, true)
@@ -105,6 +109,14 @@ func (c *Client) DryRunPolicy(mode PolicyMode, policyID string, policy io.Reader
 
 // FetchPolicy creates a request to fetch policy from the system
 func (c *Client) FetchPolicy(policyID string, returnJSON bool, policyTreeDepth uint, sizeLimit uint) ([]byte, error) {
+	if isConjurCloudURL(c.config.ApplianceURL) {
+		return nil, errors.New("Policy Fetch is not supported in Conjur Cloud")
+	}
+	err := c.VerifyMinServerVersion("1.21.1")
+	if err != nil {
+		return nil, fmt.Errorf("Policy Fetch is not supported in Conjur versions older than 1.21.1")
+	}
+
 	req, err := c.fetchPolicyRequest(policyID, returnJSON, policyTreeDepth, sizeLimit)
 	if err != nil {
 		return nil, err